# doc/api.yml
openapi: 3.0.0
info:
  title: Cakemix Server API
  description: Definition of Cakemix Server API
  version: 1.0.0
  contact:
    name: Wonder Wonder
    url: 'https://wonder-wonder.xyz'
    email: contact@wonder-wonder.xyz
servers:
  - url: 'http://localhost:8081/v1'
    description: Server URL for localhost development
  - url: 'https://api.cakemix.wonder-wonder.xyz/v1'
    description: Server URL for production
paths:
...
  '/team/{teamid}/member':
    parameters:
      - schema:
          type: string
        name: teamid
        in: path
        required: true
    get:
      summary: Get team member list
      parameters:
        - schema:
            type: integer
          in: query
          name: limit
          description: member limit
        - schema:
            type: integer
          in: query
          name: offset
          description: member offset
      responses:
        '200':
          description: Got team member list
          content:
            application/json:
              schema:
                type: object
                description: Member list
                properties:
                  total:
                    type: integer
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemberInfoModel'
        '404':
          description: The team is not found.
      operationId: get-team-teamid-member
      description: Get team member list.
      tags:
        - Team
    post:
      summary: Add the member into team
      operationId: post-team-teamid-member
      responses:
        '200':
          description: Added the member into the team.
        '401':
          description: Invalid token.
        '403':
          description: Adding member into the team is not allowed.
        '404':
          description: The team is not found.
      description: Add the member into the team.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInfoModel'
        description: Member info.
      tags:
        - Team
    put:
      summary: Modify member info in the team
      operationId: put-team-teamid-member
      responses:
        '200':
          description: Modified the member info.
        '401':
          description: Invalid token.
        '403':
          description: Modifying member info team is not allowed.
        '404':
          description: The team is not found.
      description: Modify member info in the team.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInfoModel'
        description: Member info.
      tags:
        - Team
    delete:
      summary: Remove the member from the team
      operationId: delete-team-teamid-member
      responses:
        '200':
          description: Removed the member from the tean.
        '401':
          description: Invalid token.
        '403':
          description: Removing member from the team is not allowed.
        '404':
          description: The team is not found.
      description: Remove the member from the team.
      parameters:
        - schema:
            type: string
          in: query
          name: uuid
          description: Removing member UUID
          required: true
      tags:
        - Team
  /team:
    post:
      summary: Create team
      operationId: post-team
      responses:
        '200':
          description: The new team is created successfully.
        '401':
          description: Invalid token.
        '409':
          description: The teamname is already used.
      tags:
        - Team
      description: Create new team.
      parameters:
        - schema:
            type: string
          in: query
          name: name
          description: Teamname
          required: true
    parameters: []
  '/team/{teamid}':
    parameters:
      - schema:
          type: string
        name: teamid
        in: path
        required: true
        description: Team UUID
    delete:
      summary: Remove team
      operationId: delete-team-teamid
      responses:
        '200':
          description: The team is removed successfully.
        '401':
          description: Invalid token.
        '403':
          description: Removing team is not allowed.
        '404':
          description: The team is not found.
      tags:
        - Team
      description: Remove the team.
...
  /search/team:
    get:
      summary: Get team list
      tags:
        - Search
      responses:
        '200':
          description: Got team list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileModel'
      operationId: get-search-team
      parameters:
        - schema:
            type: string
          in: query
          name: q
          description: search filter
        - schema:
            type: integer
          in: query
          name: limit
          description: search limit
        - schema:
            type: integer
          in: query
          name: offset
          description: search offset
  /search/user:
    get:
      summary: Get user list
      tags:
        - Search
      responses:
        '200':
          description: Got user list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileModel'
      operationId: get-search-user
      parameters:
        - schema:
            type: string
          in: query
          name: q
          description: search filter
        - schema:
            type: integer
          in: query
          name: limit
          description: search limit
        - schema:
            type: integer
          in: query
          name: offset
          description: search offset
...
components:
  schemas:
...
    ProfileModel:
      title: ProfileModel
      type: object
      description: |-
        Profile model for get and post of /profile/{name}.
        Required items are attached in child items.
      properties:
        uuid:
          type: string
          description: UUID for user/team
        name:
          type: string
          description: Username or teamname
        bio:
          type: string
          description: Biography
        icon_uri:
          type: string
          description: IconURI
          format: uri
        created_at:
          type: integer
          description: User/team creation date
        attr:
          type: string
          description: 'Attribute(Official, SysAdmin)'
        is_team:
          type: boolean
          description: Team or not
        teams:
          type: array
          description: List for joining teams
          items:
            $ref: '#/components/schemas/ProfileModel'
        lang:
          type: string
          description: Language
        is_admin:
          type: boolean
          description: Admin or not
      required:
        - uuid
        - name
        - icon_uri
        - attr
        - is_team
    MemberInfoModel:
      title: MemberInfoModel
      description: 'Member info model for /team/{teamid}/member'
      type: object
      properties:
        member:
          oneOf:
            - $ref: '#/components/schemas/ProfileModel'
            - type: string
              description: UUID
          description: Member profile or UUID
        permission:
          type: integer
          description: Permission
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
...
security:
  - JWT: []